+++ importlib_metadata/__init__.py
@@ -27,6 +27,7 @@ from ._compat import (
     ModuleNotFoundError,
     MetaPathFinder,
     email_message_from_string,
+    ensure_is_path,
     )
 from importlib import import_module
 from itertools import starmap
@@ -194,6 +195,15 @@ class Distribution:
             resolver()
             for resolver in cls._discover_resolvers()
             )
+
+    @staticmethod
+    def at(path):
+        """Return a Distribution for the indicated metadata path
+
+        :param path: a string or path-like object
+        :return: a concrete Distribution instance for the path
+        """
+        return PathDistribution(ensure_is_path(path))
 
     @staticmethod
     def _discover_resolvers():